requirements.txt
Flask
Flask-Cors
pymongo
requests
PyJWT
razorpay
setuptools